refactoring pkg/cs UT
