+++ go/pkg/cs/drkey/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
         "prefetcher.go",
         "secret_value_store.go",
         "service_store.go",
-        "testcommons.go",
     ],
     importpath = "github.com/scionproto/scion/go/pkg/cs/drkey",
     visibility = ["//visibility:public"],
@@ -24,13 +23,18 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = ["store_test.go"],
+    srcs = [
+        "export_test.go",
+        "secret_value_store_test.go",
+        "service_store_test.go",
+    ],
     embed = [":go_default_library"],
     deps = [
         "//go/lib/addr:go_default_library",
         "//go/lib/drkey:go_default_library",
         "//go/lib/util:go_default_library",
         "//go/lib/xtest:go_default_library",
+        "//go/pkg/cs/drkey/test:go_default_library",
         "@com_github_stretchr_testify//require:go_default_library",
     ],
 )
