+++ go/pkg/cs/drkey/export_test.go
@@ -1,0 +1,35 @@
+// Copyright 2020 ETH Zurich
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package drkey
+
+import (
+	"time"
+
+	"github.com/scionproto/scion/go/lib/drkey"
+)
+
+func (c *SecretValueStore) SetTimeNowFunction(f func() time.Time) {
+	c.mutex.Lock()
+	c.timeNowFcn = f
+	c.mutex.Unlock()
+}
+
+func (c *SecretValueStore) CleanExpired() {
+	c.cleanExpired()
+}
+
+func (c *SecretValueStore) Cache() map[int64]drkey.SV {
+	return c.cache
+}
