+++ go/pkg/cs/drkey/grpc/BUILD.bazel
@@ -19,6 +19,7 @@ go_library(
         "//go/lib/log:go_default_library",
         "//go/lib/scrypto/cppki:go_default_library",
         "//go/lib/serrors:go_default_library",
+        "//go/pkg/cs/drkey:go_default_library",
         "//go/pkg/grpc:go_default_library",
         "//go/pkg/proto/control_plane:go_default_library",
         "//go/pkg/proto/drkey:go_default_library",
@@ -45,7 +46,7 @@ go_test(
         "//go/lib/scrypto/cppki:go_default_library",
         "//go/lib/util:go_default_library",
         "//go/lib/xtest:go_default_library",
-        "//go/pkg/cs/drkey:go_default_library",
+        "//go/pkg/cs/drkey/test:go_default_library",
         "//go/pkg/proto/control_plane:go_default_library",
         "//go/pkg/trust:go_default_library",
         "//go/pkg/trust/mock_trust:go_default_library",
